requirements.txt
termcolor
pyopenssl